% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census10_to_isco08.R
\name{census10_to_isco08}
\alias{census10_to_isco08}
\title{Convert Census 2010 to ISCO08}
\usage{
census10_to_isco08(data, census_col, one_to_one = FALSE)
}
\arguments{
\item{data}{A data frame containing Census 2010 codes.}

\item{census_col}{The column name containing Census 2010 codes.}

\item{one_to_one}{Logical. If TRUE, selects the closest match by title.}
}
\value{
A data frame with the ISCO08 mappings.
}
\description{
Maps Census 2010 occupational codes to ISCO08 occupational codes using a predefined crosswalk.
}
