% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{crosswalk_census2010_isco08}
\alias{crosswalk_census2010_isco08}
\title{Crosswalk Dataset for Census 2010 to ISCO08}
\format{
A data frame with the following columns:
\describe{
\item{occ2010}{Census 2010 occupational codes.}
\item{occ2010title}{Titles for Census 2010 occupational codes.}
\item{isco08}{ISCO08 occupational codes.}
\item{iscolabel}{Titles for ISCO08 occupational codes.}
}
}
\source{
US BLS
}
\usage{
crosswalk_census2010_isco08
}
\description{
A crosswalk dataset mapping Census 2010 occupational codes to ISCO08.
}
\keyword{datasets}
